Added Tests for Modified Following and Preceeding... you can use them for JDK... BUT I'm not sure it is 100%.
For instance at the moment, the date moves FIRST by the number of days given and it handles the holiday only for that calculated date.  So it would be possible to jump to another month e.g. 31/7/06 + 1 day or 30/7/06 +5 days...

i do not think it is wrong, need to be checked